@@ -1,4 +1,5 @@
 /*
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  * Copyright (c) 1997, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
@@ -33,8 +34,6 @@ import java.util.jar.JarFile;
 import java.util.jar.JarOutputStream;
 import java.util.zip.ZipException;
 
-import org.glassfish.main.jdke.i18n.LocalStringsImpl;
-
 import static java.util.jar.Attributes.Name.CLASS_PATH;
 
 /**
@@ -68,7 +67,6 @@ public class PackageAppClient {
     private final static String GLASSFISH_LIB = "glassfish/lib";
     private final static String GLASSFISH_BIN = "glassfish/bin";
     private final static String GLASSFISH_CONFIG = "glassfish/config";
-    private final static String MODULES_ENDORSED_DIR = "glassfish/modules/endorsed";
     private final static String MQ_LIB = "mq/lib";
     private final static String DOMAIN_1_CONFIG = "glassfish/domains/domain1/config";
     private final static String INDENT = "  ";
@@ -79,15 +77,9 @@ public class PackageAppClient {
     private final static String[] DIRS_TO_COPY = new String[] {
         GLASSFISH_LIB + "/dtds",
         GLASSFISH_LIB + "/schemas",
-        GLASSFISH_LIB + "/appclient" };
-
-    /*
-     * relative path to the endorsed directory of the app server. Handled separately from other directorys because we do not
-     * include all files from the endorsed directory.
-     */
-    private final static String LIB_ENDORSED_DIR = GLASSFISH_LIB + "/endorsed";
-
-    private final static String[] ENDORSED_DIRS_TO_COPY = new String[] { LIB_ENDORSED_DIR, MODULES_ENDORSED_DIR };
+        GLASSFISH_LIB + "/appclient",
+        GLASSFISH_LIB + "/bootstrap",
+        };
 
     /* default sun-acc.xml is relative to the installation directory */
     private final static String DEFAULT_ACC_XML = DOMAIN_1_CONFIG + ACC_CONFIG_FILE_DEFAULT;
@@ -124,8 +116,6 @@ public class PackageAppClient {
     /* default output file */
     private final static String DEFAULT_OUTPUT_PATH = GLASSFISH_LIB + "/appclient.jar";
 
-    private final static LocalStringsImpl strings = new LocalStringsImpl(PackageAppClient.class);
-
     private boolean isVerbose;
 
     /**
@@ -146,74 +136,62 @@ public class PackageAppClient {
         File installDir = findInstallDir(thisJarFile);
         File modulesDir = new File(installDir.toURI().resolve("glassfish/modules/"));
 
-        /*
-         * Write the new JAR to a temp file in the install directory. Then we can simply rename the file to the correct name.
-         * (Rename does not work on Windows systems across volumes.)
-         */
+        // Write the new JAR to a temp file in the install directory. Then we can simply rename
+        // the file to the correct name.
+        // (Rename does not work on Windows systems across volumes.)
         File tempFile = File.createTempFile("appc", ".tmp", installDir);
         File outputFile = chooseOutputFile(installDir, args);
 
         File[] configFiles = chooseConfigFiles(installDir, args);
         String[] classPathElements = getJarClassPath(thisJarFile).split(" ");
 
-        JarOutputStream os = new JarOutputStream(new BufferedOutputStream(new FileOutputStream(tempFile)));
+        try (JarOutputStream os = new JarOutputStream(new BufferedOutputStream(new FileOutputStream(tempFile)))) {
+            // Add this JAR file to the output.
+            addFile(os, installDir.toURI(), thisJarFile.toURI(), tempFile, "");
 
-        /*
-         * Add this JAR file to the output.
-         */
-        addFile(os, installDir.toURI(), thisJarFile.toURI(), tempFile, "");
+            // JARs listed in the Class-Path are all relative to the modules directory so resolve
+            // each Class-Path entry against the modules directory.
+            for (String classPathElement : classPathElements) {
+                File classPathJAR = new File(modulesDir, classPathElement);
+                addFile(os, installDir.toURI(), modulesDir.toURI().resolve(classPathJAR.toURI()), tempFile, "");
+            }
 
-        /*
-         * JARs listed in the Class-Path are all relative to the modules directory so resolve each Class-Path entry against the
-         * modules directory.
-         */
-        for (String classPathElement : classPathElements) {
-            File classPathJAR = new File(modulesDir, classPathElement);
-            addFile(os, installDir.toURI(), modulesDir.toURI().resolve(classPathJAR.toURI()), tempFile, "");
+            // The directories to copy are all relative to the installation directory, so resolve
+            // them against the installDir file.
+            for (String dirToCopy : DIRS_TO_COPY) {
+                addDir(os, installDir.toURI(), installDir.toURI().resolve(dirToCopy), tempFile, "");
+            }
+
+            for (String singleFileToCopy : SINGLE_FILES_TO_COPY) {
+                addFile(os, installDir.toURI(), installDir.toURI().resolve(singleFileToCopy), tempFile, "");
+            }
+
+            // The glassfish-acc.xml file and sun-acc.xml files.
+            for (File configFile : configFiles) {
+                addFile(os, installDir.toURI(), configFile.toURI(), tempFile, "");
+            }
         }
-
-        /*
-         * The directories to copy are all relative to the installation directory, so resolve them against the installDir file.
-         */
-        for (String dirToCopy : DIRS_TO_COPY) {
-            addDir(os, installDir.toURI(), installDir.toURI().resolve(dirToCopy), tempFile, "");
-        }
-
-        for (String endorsedDirToCopy : ENDORSED_DIRS_TO_COPY) {
-            addEndorsedFiles(os, installDir.toURI(), installDir.toURI().resolve(endorsedDirToCopy), tempFile);
-        }
-
-        for (String singleFileToCopy : SINGLE_FILES_TO_COPY) {
-            addFile(os, installDir.toURI(), installDir.toURI().resolve(singleFileToCopy), tempFile, "");
-        }
-
-        /*
-         * The glassfish-acc.xml file and sun-acc.xml files.
-         */
-        for (File configFile : configFiles) {
-            addFile(os, installDir.toURI(), configFile.toURI(), tempFile, "");
-        }
-
-        os.close();
         placeFile(tempFile, outputFile);
     }
 
     private void placeFile(final File tempFile, final File outputFile) {
         if (outputFile.exists()) {
             if (!outputFile.delete()) {
-                throw new RuntimeException(strings.get("errDel", outputFile.getAbsolutePath()));
+                throw new RuntimeException("Stopping; could not delete the existing output file " + outputFile.getAbsolutePath());
             }
-            System.out.println(strings.get("replacingFile", outputFile.getAbsolutePath()));
+            System.out.println("Replacing " + outputFile.getAbsolutePath());
         } else {
-            System.out.println(strings.get("creatingFile", outputFile.getAbsolutePath()));
+            System.out.println("Creating" + outputFile.getAbsolutePath());
         }
 
         if (isVerbose) {
-            System.out.println(strings.get("moving", tempFile.getAbsolutePath(), outputFile.getAbsolutePath()));
+            System.out
+                .println("Moving temp file " + tempFile.getAbsolutePath() + " to " + outputFile.getAbsolutePath());
         }
 
         if (!tempFile.renameTo(outputFile)) {
-            throw new RuntimeException(strings.get("errRenaming", tempFile.getAbsolutePath(), outputFile.getAbsolutePath()));
+            throw new RuntimeException(
+                "Error renaming temp file " + tempFile.getAbsolutePath() + " to " + outputFile.getAbsolutePath());
         }
     }
 
@@ -243,7 +221,7 @@ public class PackageAppClient {
     private void addFile(JarOutputStream os, URI installDirURI, URI absoluteURIToAdd, File outputFile, String indent) throws IOException {
         try {
             if (isVerbose) {
-                System.err.println(indent + strings.get("addingFile", absoluteURIToAdd));
+                System.err.println(indent + "Adding directory " + absoluteURIToAdd);
             }
 
             File fileToCopy = new File(absoluteURIToAdd);
@@ -261,7 +239,8 @@ public class PackageAppClient {
                  */
                 if (!new File(absoluteURIToAdd).exists()) {
                     if (isVerbose) {
-                        System.err.println(indent + strings.get("noFile", new File(absoluteURIToAdd).getAbsolutePath()));
+                        System.err.println(indent + "Error locating file "
+                            + new File(absoluteURIToAdd).getAbsolutePath() + "; continuing");
                     }
                     return;
                 }
@@ -276,7 +255,8 @@ public class PackageAppClient {
                  * Probably duplicate entry. Keep going after logging the error.
                  */
                 if (isVerbose) {
-                    System.err.println(indent + strings.get("zipExc", e.getLocalizedMessage()));
+                    System.err.println(
+                        indent + "Continuing after ZipException when adding a file: " + e.getLocalizedMessage());
                 }
             } catch (FileNotFoundException ignore) {
             }
@@ -323,7 +303,7 @@ public class PackageAppClient {
         }
 
         if (isVerbose) {
-            System.err.println(indent + strings.get("addingDir", dirFile.getAbsolutePath()));
+            System.err.println(indent + "Adding dir " + dirFile.getAbsolutePath());
         }
 
         for (File fileToAdd : matchingFiles) {
@@ -406,7 +386,8 @@ public class PackageAppClient {
             File userSpecifiedFile = new File(xmlArg);
             files = new File[] { userSpecifiedFile };
             if (!userSpecifiedFile.exists()) {
-                System.err.println(strings.get("xmlNotFound", userSpecifiedFile.getAbsolutePath()));
+                System.err.println("The XML configuration file " + userSpecifiedFile.getAbsolutePath()
+                    + " does not exist; continuing but output is incomplete");
             }
 
         }
@@ -414,7 +395,7 @@ public class PackageAppClient {
 
     }
 
-    private File findInstallDir(File currentJarFile) throws URISyntaxException {
+    private File findInstallDir(File currentJarFile) {
         return currentJarFile.getParentFile().getParentFile().getParentFile();
     }
 
