@@ -1,4 +1,5 @@
 /*
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  * Copyright (c) 1997, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
@@ -33,9 +34,9 @@ import java.util.logging.Logger;
 
 import javax.swing.SwingUtilities;
 
-import org.glassfish.appclient.client.acc.UserError;
 import org.glassfish.appclient.client.jws.boot.ErrorDisplayDialog;
 import org.glassfish.appclient.client.jws.boot.LaunchSecurityHelper;
+import org.glassfish.embeddable.client.UserError;
 
 /**
  *
@@ -88,18 +89,14 @@ public class JWSAppClientContainerMain {
 
         try {
             now = System.currentTimeMillis();
-            /*
-             * Process any arguments (conveyed as properties in the JNLP) intended for the JWS-aware ACC.
-             */
+            // Process any arguments (conveyed as properties in the JNLP) intended for the JWS-aware ACC.
             processJWSArgs();
 
             final String agentArgsText = System.getProperty("agent.args");
             LaunchSecurityHelper.setPermissions();
 
-            /*
-             * Prevent the Java Web Start class loader from delegating to its parent when resolving classes and resources that
-             * should come from the GlassFish-provided endorsed JARs.
-             */
+            // Prevent the Java Web Start class loader from delegating to its parent when resolving
+            // classes and resources that should come from the GlassFish-provided endorsed JARs.
             insertMaskingLoader();
 
             final ClientRunner runner = new ClientRunner(agentArgsText, args);
@@ -133,8 +130,8 @@ public class JWSAppClientContainerMain {
         @Override
         public void run() {
             try {
-                AppClientFacade.prepareACC(agentArgsText, null);
-                AppClientFacade.launch(args);
+                AppClientContainerHolder.init(agentArgsText, null);
+                AppClientContainerHolder.getInstance().launch(args);
                 logger.log(Level.FINE, "JWSAppClientContainer finished after {0} ms", (System.currentTimeMillis() - now));
             } catch (UserError ue) {
                 if (!isTestMode()) {
@@ -155,9 +152,7 @@ public class JWSAppClientContainerMain {
         props.load(sr);
 
         final ClassLoader jwsLoader = Thread.currentThread().getContextClassLoader();
-
         final ClassLoader mcl = getMaskingClassLoader(jwsLoader.getParent(), props);
-
         final Field jwsLoaderParentField = ClassLoader.class.getDeclaredField("parent");
         jwsLoaderParentField.setAccessible(true);
         jwsLoaderParentField.set(jwsLoader, mcl);
