@@ -34,6 +34,7 @@ import java.security.PrivilegedAction;
 import javax.xml.stream.XMLStreamException;
 
 import org.glassfish.api.deployment.archive.ReadableArchive;
+import org.glassfish.embeddable.client.UserError;
 import org.glassfish.hk2.api.ServiceLocator;
 import org.xml.sax.SAXException;
 
@@ -104,9 +105,9 @@ interface Launchable {
                  * ACCClassLoader instance yet.  Create one if needed
                  * before proceeding.
                  */
-                ACCClassLoader cl = ACCClassLoader.instance();
+                TransformingClassLoader cl = TransformingClassLoader.instance();
                 if (cl == null) {
-                    cl = ACCClassLoader.newInstance(Thread.currentThread().getContextClassLoader(), false);
+                    cl = TransformingClassLoader.newInstance(Thread.currentThread().getContextClassLoader(), false);
                 }
                 cl.appendURL(clientOrFacadeURL);
             }
@@ -122,15 +123,8 @@ interface Launchable {
                 final ReadableArchive facadeRA,
                 final ReadableArchive clientRA) throws IOException, SAXException {
             archivist.setAnnotationProcessingRequested(true);
-            final ACCClassLoader tempLoader = AccessController.doPrivileged(
-                    new PrivilegedAction<ACCClassLoader>() {
-
-                        @Override
-                        public ACCClassLoader run() {
-                            return new ACCClassLoader(loader.getURLs(), loader.getParent());
-                        }
-                    }
-                );
+            PrivilegedAction<TransformingClassLoader> action = () -> new TransformingClassLoader(loader.getURLs(), loader.getParent());
+            final TransformingClassLoader tempLoader = AccessController.doPrivileged(action);
             archivist.setClassLoader(tempLoader);
 
             final ApplicationClientDescriptor acDesc = archivist.open(facadeRA, clientRA);
