@@ -83,15 +83,15 @@ public class ProgramOptions {
     private PasswordLocation location;
     private String commandName;
 
-    /*
+
+    private String modulePath;
+    /**
      * Information passed in from AsadminMain and used by start-domain.
-     * XXX - this is somewhat of a kludge but this seems the best place
-     * to put it for now
      */
     private String classPath;
     private String className;
 
-    /*
+    /**
      * Define the meta-options known by the asadmin command.
      */
     static {
@@ -153,6 +153,7 @@ public class ProgramOptions {
         this.options = new ParameterMap(other.options);
         this.env = other.env;
         this.password = other.password;
+        this.modulePath = other.modulePath;
         this.classPath = other.classPath;
         this.className = other.className;
     }
@@ -638,6 +639,20 @@ public class ProgramOptions {
      */
     public String getPlainOption(String name) {
         return options.getOne(name);
+    }
+
+    /**
+     * @return the modulePath
+     */
+    public String getModulePath() {
+        return modulePath;
+    }
+
+    /**
+     * @param modulePath the modulePath to set
+     */
+    public void setModulePath(String modulePath) {
+        this.modulePath = modulePath;
     }
 
     /**
