@@ -1,4 +1,5 @@
 /*
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  * Copyright (c) 1997, 2018 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
@@ -14,33 +15,36 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.appclient.client.acc;
+package org.glassfish.embeddable.client;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import java.text.MessageFormat;
 
 /**
- *Represents any user error, such as an invalid combination of command options,
- *or specifying a non-existent JAR file.
- *<p>
- *Such errors should be user-correctable, provided the message is clear.
- *So no stack traces should be displayed when UserErrors are thrown.
+ * Represents any user error, such as an invalid combination of command options,
+ * or specifying a non-existent JAR file.
+ * <p>
+ * Such errors should be user-correctable, provided the message is clear.
+ * So no stack traces should be displayed when UserErrors are thrown.
  *
  * @author tjquinn
  */
 public class UserError extends Throwable {
 
+    private static final long serialVersionUID = 4171944609382558583L;
+
     /** Allows user to turn on stack traces for user errors - normally off */
-    private static final String SHOW_STACK_TRACES_PROPERTY_NAME =
-            UserError.class.getPackage().getName() + ".showUserErrorStackTraces";
+    private static final String SHOW_STACK_TRACES_PROPERTY_NAME = UserError.class.getPackage().getName()
+        + ".showUserErrorStackTraces";
 
     /**
      * Creates a new UserError instance having formatted the message with the
      * arguments provided.
-     *@param message the message string, presumably containing argument placeholders
-     *@param args 0 or more arguments for substitution for the placeholders in the message string
-     *@return new UserError with message formatted as requested
+     *
+     * @param message the message string, presumably containing argument placeholders
+     * @param args 0 or more arguments for substitution for the placeholders in the message string
+     * @return new UserError with message formatted as requested
      */
     public static UserError formatUserError(String message, String... args) {
         String formattedMessage = MessageFormat.format(message, (Object[]) args);
@@ -49,7 +53,7 @@ public class UserError extends Throwable {
     }
 
     /** xmlMessage implementation showed the usage message after the error */
-    private String usage = null;
+    private String usage;
 
     public UserError(String message) {
         super(message);
@@ -63,24 +67,27 @@ public class UserError extends Throwable {
         super(cause);
     }
 
+
     /**
-     *Sets whether or not the usage message should be displayed after the
-     *error message is displayed to the user.
-     *@param showUsage the new setting
+     * Sets whether or not the usage message should be displayed after the
+     * error message is displayed to the user.
+     *
+     * @param usage the new setting
      */
     public void setUsage(String usage) {
         this.usage = usage;
     }
 
+
     /**
-     *Displays the user error message, and any messages along the exception
-     *chain, if any, and then exits.  If showUsage has been set to true, then
-     *the usage message is displayed before exiting.
-     *<p>
-     *Only the messages, and not the stack traces, are shown because these are
-     *user errors that should be user-correctable.  Stack traces are too
-     *alarming and of minimal use to the user as he or she tries to understand
-     *and fix the error.
+     * Displays the user error message, and any messages along the exception
+     * chain, if any, and then exits. If showUsage has been set to true, then
+     * the usage message is displayed before exiting.
+     * <p>
+     * Only the messages, and not the stack traces, are shown because these are
+     * user errors that should be user-correctable. Stack traces are too
+     * alarming and of minimal use to the user as he or she tries to understand
+     * and fix the error.
      */
     public void displayAndExit() {
         display(System.err);
