@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, 2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2022, 2025 Contributors to the Eclipse Foundation
  * Copyright (c) 2010, 2018 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
@@ -51,11 +51,13 @@ public class RestartServer {
     private Properties props;
     private Logger logger;
     private boolean verbose;
+    private String modulepath;
     private String classpath;
     private String classname;
     private String argsString;
     private String[] args;
     private String serverName = "";
+
     private static final LocalStringsImpl strings = new LocalStringsImpl(RestartServer.class);
     private static final String AS_RESTART_PID = "-DAS_RESTART=" + ProcessHandle.current().pid();
     private static final String[] normalProps = { AS_RESTART_PID };
@@ -134,13 +136,15 @@ public class RestartServer {
 
     private void scheduleReincarnation() throws RDCException {
         try {
-            Runtime.getRuntime().addShutdownHook(new StartServerShutdownHook(classpath, normalProps, classname, args));
+            Runtime.getRuntime()
+                .addShutdownHook(new StartServerShutdownHook(modulepath, classpath, normalProps, classname, args));
         } catch (Exception e) {
             throw new RDCException(e);
         }
     }
 
     private boolean setupReincarnationWithAsadmin() throws RDCException {
+        modulepath = props.getProperty(BootstrapKeys.ASADMIN_MP);
         classpath = props.getProperty(BootstrapKeys.ASADMIN_CP);
         classname = props.getProperty(BootstrapKeys.ASADMIN_CN);
         argsString = props.getProperty(BootstrapKeys.ASADMIN_ARGS);
@@ -149,7 +153,7 @@ public class RestartServer {
     }
 
     private boolean setupReincarnationWithOther() throws RDCException {
-
+        modulepath = props.getProperty(BootstrapKeys.ORIGINAL_MP);
         classpath = props.getProperty(BootstrapKeys.ORIGINAL_CP);
         classname = props.getProperty(BootstrapKeys.ORIGINAL_CN);
         argsString = props.getProperty(BootstrapKeys.ORIGINAL_ARGS);
@@ -163,12 +167,12 @@ public class RestartServer {
         // 1) true
         // 2) false
         // 3) RDCException
-        if (classpath == null && classname == null && argsString == null) {
+        if (classpath == null && modulepath == null && classname == null && argsString == null) {
             return false;
         }
 
         // now that at least one is set -- demand that ALL OF THEM be set...
-        if (!ok(classpath) || !ok(classname) || argsString == null) {
+        if (!ok(classpath) || !ok(modulepath) || !ok(classname) || argsString == null) {
             throw new RDCException(strings.get(errorStringKey));
         }
 
@@ -265,7 +269,7 @@ public class RestartServer {
     }
 
     private boolean ok(String s) {
-        return s != null && s.length() > 0;
+        return s != null && !s.isEmpty();
     }
 
     // We use this simply to tell the difference between fatal errors and other
