@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Contributors to the Eclipse Foundation.
+ * Copyright (c) 2024, 2025 Contributors to the Eclipse Foundation.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -56,8 +56,6 @@ class StartServerShutdownHook extends Thread {
 
     private static final Logger LOG = System.getLogger(StartServerShutdownHook.class.getName());
     private static final boolean LOG_RESTART = Boolean.parseBoolean(System.getenv("AS_RESTART_LOGFILES"));
-    private static final Path CFGDIR = new File(System.getProperty("com.sun.aas.instanceRoot"), "config").toPath()
-        .toAbsolutePath();
     private static final Path LOGDIR = new File(System.getProperty("com.sun.aas.instanceRoot"), "logs").toPath()
         .toAbsolutePath();
     private static final Predicate<Thread> FILTER_OTHER_HOOKS = t -> t.getName().startsWith("GlassFish")
@@ -67,7 +65,7 @@ class StartServerShutdownHook extends Thread {
     private final ProcessBuilder builder;
     private final Instant startTime;
 
-    StartServerShutdownHook(String classpath, String[] sysProps, String classname, String[] args) {
+    StartServerShutdownHook(String modulepath, String classpath, String[] sysProps, String classname, String[] args) {
         super("GlassFish Restart Shutdown Hook");
         setDaemon(false);
         if (classname == null || classname.isBlank()) {
@@ -75,7 +73,7 @@ class StartServerShutdownHook extends Thread {
         }
         this.startTime = Instant.now();
         this.logFile = getLogFileOld(startTime);
-        this.builder = prepareStartup(startTime, classpath, sysProps, classname, args);
+        this.builder = prepareStartup(startTime, modulepath, classpath, sysProps, classname, args);
     }
 
 
@@ -190,16 +188,27 @@ class StartServerShutdownHook extends Thread {
     }
 
 
-    private static ProcessBuilder prepareStartup(Instant startTime, String classpath, String[] sysprops, String classname,
-        String[] args) {
-        final Path javaExecutable = detectJavaExecutable();
+    private static ProcessBuilder prepareStartup(Instant startTime, String modulepath, String classpath,
+        String[] sysprops, String classname, String[] args) {
+        final File javaExecutable = detectJavaExecutable().toFile();
+        if (!javaExecutable.canExecute()) {
+            throw new Error(javaExecutable + " is not an existing executable file.");
+        }
         final List<String> cmdline = new ArrayList<>();
         if (!OS.isWindows()) {
             cmdline.add("nohup");
         }
-        cmdline.add(javaExecutable.toFile().getAbsolutePath());
-        cmdline.add("-cp");
-        cmdline.add(classpath);
+        cmdline.add(javaExecutable.getAbsolutePath());
+        if (modulepath != null && !modulepath.isEmpty()) {
+            cmdline.add("--module-path");
+            cmdline.add(modulepath);
+            cmdline.add("--add-modules");
+            cmdline.add("ALL-MODULE-PATH");
+        }
+        if (classpath != null && !classpath.isEmpty()) {
+            cmdline.add("-cp");
+            cmdline.add(classpath);
+        }
         if (sysprops != null) {
             for (String sysprop : sysprops) {
                 cmdline.add(sysprop);
